// src/types.ts
export enum ContentCategory {
    HEADER = "HEADER",
    FOOTER = "FOOTER",
    HERO = "HERO",
    ABOUT = "ABOUT",
    SERVICES = "SERVICES",
    TEXT = "TEXT",
    CONTACT = "CONTACT"
}

export interface ProjectStruct {
    id: string
    title: string
    pages: Record<string, PageStruct>
    colourScheme: ColourScheme
    fontScheme: FontScheme
}

export interface PageStruct {
    id: string
    path: string
    description: string
    content: ContentStruct[]
    contentActive: ContentStruct | null
    colourScheme: ColourScheme
    fontScheme: FontScheme
}

export interface ContentStruct {
    id: string
    description: string
    Component: React.FC | string
    categories: ContentCategory[]
    props: any
}
// ...
export enum PageContentMessageType {
    SET_ID = "SET_ID",
    SET_DESCRIPTION = "SET_DESCRIPTION",
    SET_COMPONENT = "SET_COMPONENT",
    SET_PROPS = "SET_PROPS",
    SET_CATEGORIES = "SET_CATEGORIES",
}

export type PageContentEvent = {
    type: PageContentMessageType.SET_ID
    data: string
} | {
    type: PageContentMessageType.SET_DESCRIPTION
    data: string
} | {
    type: PageContentMessageType.SET_COMPONENT
    data: React.FC
} | {
    type: PageContentMessageType.SET_PROPS
    data: unknown
} | {
    type: PageContentMessageType.SET_CATEGORIES
    data: ContentCategory[]
}